Make LP64-clean, fix a possible buffer overrun, and link errors.
There were a few string parsing routines that were confusingly coercing
character pointers to integers when attempting to do pointer math. This
makes LP64 systems very angry and just isn't necessary.

Fixed an issue in the callbook search routines that could cause a potential
buffer overflow, which was kindly pointed out by clang.

Finally, there were a few places where several modules within the same program
were duplicating definitions of global variables that they share. While this
may have been ok under the old Sun tools, and even under older versions of
GCC/binutils, it is not ok with the latest linker in FreeBSD 13. The fix
is to decisively define all inter-module globals in main.c and use
`extern` declarations to reference them outside of main.

+++ src/bbs/gated/guess.c
@@ -80,7 +80,7 @@ disp_guess_on_domain(char *s)
 
 	if((p = (char*)rindex(s, '.')) != NULL) {
 		p--;
-		while((int)p > (int)s) {
+		while(p > s) {
 			if(*p == '.') {
 				p++;
 #if HAVE_REGCOMP
